remove all DSE code since that was causing TSC issues (#1150)
This code was causing `tsc` to reach a stack call limit. Removing it
made `tsc` work so I think the webpack issues may be related.

This is old code for a feature we never released due to legal issues.
We're not working to resolve those issues or pushing this feature
forward, so I'm making the call that it's safe to delete this code!

still need to double check that this works well in the application!