fix bug when switching threads (#1155)
## Description

There's an issue in prod when switching between existing threads. Loader
is short circuited when data is already cached for a thread in the
queryClient. This results in model not being updated when a cached
thread is loaded, it just loads the default model.

This doesn't make sense to me why the load isn't allowed to run. 

## Reviewer Instructions

I ran single and comparison models and they seem to work fine without
this code.

Please manually test

## Testing Plan

tested manually