don't throw error in token check (#1162)
## Description

### Issue
Refresh token error being displayed.

### Resolution

Per @mtblanton  -- we can remove the refresh token all together.

closes: allenai/playground-issues-repo#511

## Reviewer Instructions

I've also removed the re-throwing, as this is what caused the flash of
error. I think if we plan on handling it, we should probably not throw
also (as this gets handled by react-router).

I was on the fence of keeping the error throwing and handling it in
playgroundApiClient middleware.

## Testing Plan

Describe what you did to test the change.

## Output / Screenshots

### Before

What did it look like before?

### After

What will it look like after merging?

### A11y

How do keyboard navigation and screen readers work with this PR?