Infini-gram index in model config (#1170)
## Description

Closes https://github.com/allenai/playground-issues-repo/issues/576

The biggest change in this PR is how the API schema is generated. I
changed it so we get the values of enums in an array. This required a
change from a .d.ts file to a .ts file for the schema so we could use it
in other files.

I also tried to clean up this part of the repo a bit by throwing things
into folders.

## Testing Plan

Tested locally through my local API.

## Output / Screenshots

<img width="202" height="247" alt="image"
src="https://github.com/user-attachments/assets/f74db3c9-a5ae-4402-bd56-0ba6530b470b"
/>