Hide olmo trace on models that don't have an index (#1177)
## Description

Hides the button if model or model.infinigram index is missing. This PR
relies on the types added on
https://github.com/allenai/olmo-api/pull/380/ and fixes the error caused
by the same PR.

## Reviewer Instructions

Anything a review needs to know about this PR to help them give better
feedback. Are there specific questions you have? Are there areas you are
unsure of? Does this come from or relate to another part of the
codebase?

## Testing Plan

Check a model with or without an index. Note: you might need to update
your database with indexes or this won't show for you.

## Output / Screenshots

### After

<img width="688" height="683" alt="Screenshot 2025-09-12 at 8 50 52 AM"
src="https://github.com/user-attachments/assets/62fe63a8-00ff-49fb-b85d-e686a14d1a16"
/>

### A11y

N/A