Add defaultExpanded to ToolCallWidget - set to true for user tool cals, false otherwise (#1179)
Closes: https://github.com/allenai/playground-issues-repo/issues/614

## Description

Add a `defaultExpanded` to the ToolCallWidget, have it default to `true`
for `'user_defined'` and `false` otherwise

## Reviewer Instructions

Anything a review needs to know about this PR to help them give better
feedback. Are there specific questions you have? Are there areas you are
unsure of? Does this come from or relate to another part of the
codebase?

## Testing Plan

Describe what you did to test the change.

## Output / Screenshots

### Before

What did it look like before?

### After

What will it look like after merging?

### A11y

How do keyboard navigation and screen readers work with this PR?