Fix hasPoints to use search vs test (#1194)
## Description

When using Molmo and pointing, switching back and forth between raw
loses the reply image with points.

hasPoint is using a global (`/g`) regex and `RegExp.test()`, which will
search from the last match (when global).

This causes a second call to hasPoints to fail when the first succeeded.

Changing this to `String.search(regex) !== -1` fixes this behavior

## Reviewer Instructions

## Testing Plan

Describe what you did to test the change.

## Output / Screenshots

### Before


https://github.com/user-attachments/assets/1459a0f5-ab90-4c19-a2ef-c0ffb0df09b4



### After



https://github.com/user-attachments/assets/85e4479c-6174-4472-8911-f447cf7e6f83



### A11y

How do keyboard navigation and screen readers work with this PR?