Fix duplicate thread add (#1197)
## Description
This PR fixes a bug where thread would get added twice when a tool is
called. I'm opting to change the logic to use the end token to trigger
the thread add and the form clear.

## Reviewer Instructions

Use a model with tool calling to call a tool and check the thread drawer
after the response is streamed in.

### A11y

N/A