revalidate the route when hitting /admin -- this prevents blank admin page (#1201)
This is bugging me, I had it in some code from the auth fix, but it
didn't make it into the final version.

This makes it so that clicking admin from within the admin area doesn't
give you that blank screen.

If we decide to respond to /admin with a non 3xx request, we can remove
this code.

## Description

Describe the change and its purpose. Include a ticket if applicable.

## Reviewer Instructions

Anything a review needs to know about this PR to help them give better
feedback. Are there specific questions you have? Are there areas you are
unsure of? Does this come from or relate to another part of the
codebase?

## Testing Plan

Describe what you did to test the change.

## Output / Screenshots

### Before

What did it look like before?

### After

What will it look like after merging?

### A11y

How do keyboard navigation and screen readers work with this PR?