Tools in placeholder (#1205)
## Description

This PR adds a tool calling component to placeholder page, which allows
us to bring attention to tool calling in models.

The tools are now disabled to begin with.

I made the modal/dialog scroll contents correctly and made it more
responsive (its not perfect, but much better).

## Reviewer Instructions


## Testing Plan

Describe what you did to test the change.

## Output / Screenshots

### After


https://github.com/user-attachments/assets/f79643d0-5ffa-4f07-8606-b3eabb92e4b2

### A11y

How do keyboard navigation and screen readers work with this PR?