clip value before passing to onChange (#1211)
## Description

There is a small bug where the value gets forwarded to the onChange
handler unclipped (it does clip before saving internally in the
component), this fixes that.

## Reviewer Instructions

Anything a review needs to know about this PR to help them give better
feedback. Are there specific questions you have? Are there areas you are
unsure of? Does this come from or relate to another part of the
codebase?

## Testing Plan

Describe what you did to test the change.

## Output / Screenshots

### Before

What did it look like before?

### After

What will it look like after merging?

### A11y

How do keyboard navigation and screen readers work with this PR?