better tool call result output (#1215)
I was in this code and this was bugging me, I pulled these two things
out into another PR.

## Description

Added break word css prop, which is a stricter breaking value than we
had.

Added margin to the intersection observer hook, to ensure it triggers
(it seemed to most of the time, this is more reliable).

## Reviewer Instructions

Anything a review needs to know about this PR to help them give better
feedback. Are there specific questions you have? Are there areas you are
unsure of? Does this come from or relate to another part of the
codebase?

## Testing Plan

Describe what you did to test the change.

## Output / Screenshots

### Before

What did it look like before?


https://github.com/user-attachments/assets/b9b3bfce-2be3-4ac9-a3c7-e2976c28d434


### After



https://github.com/user-attachments/assets/3dc815dd-366f-46dc-926b-f73784b8f64d



### A11y

How do keyboard navigation and screen readers work with this PR?