Make the Agent Chat page work (#1207)
## Description

Closes https://github.com/allenai/playground-issues-repo/issues/698

This PR makes the "Agent chat" page functional with a mock. It adds a
new `AgentChatProvider` that handles things specific to agent chats. It
also breaks parts of `SingleThreadProvider` out to hooks to make things
easier to read and maintain.


https://github.com/user-attachments/assets/b51d8087-abd1-4609-bd8e-fc50f0c6c155

## Reviewer Instructions

This is a separate context for just Agent chat threads. They won't
currently support comparison, we'll have to add that later.

I started looking into changing the `SingleThreadContext` to support
agent threads but it would've been messier than a new context IMO.

I'd like to separate some of these properties into different contexts so
we can reuse more but that level of refactoring is past what I want to
do on this deadline.


## Testing Plan

Tested locally with mocks. The APIs will likely change so I didn't want
to dive too deep into tests right now.

---------

Co-authored-by: Byron Bischoff <byronb@gmail.com>
Co-authored-by: Byron <bronzle@users.noreply.github.com>