fixes for contrast (#1210)
## Description

This fixes: allenai/playground-issues-repo#619

This is a quick fix for that ticket.

It has hard coded changes from Karen.
These should really be updates to varnish-theme and pull them in.

## Reviewer Instructions

Anything a review needs to know about this PR to help them give better
feedback. Are there specific questions you have? Are there areas you are
unsure of? Does this come from or relate to another part of the
codebase?

## Testing Plan

Describe what you did to test the change.

## Output / Screenshots

### Before

What did it look like before?

### After

What will it look like after merging?

### A11y

How do keyboard navigation and screen readers work with this PR?