Add the colorMode className to the body -- remove unneeded colorMode (#1217)
## Description

This moves the className to the body, which supports things like
Modals/ToolTips being portaled (we are going in and handling on a case
by case basis right now).



## Reviewer Instructions

Anything a review needs to know about this PR to help them give better
feedback. Are there specific questions you have? Are there areas you are
unsure of? Does this come from or relate to another part of the
codebase?

## Testing Plan

Describe what you did to test the change.

## Output / Screenshots

### Before

What did it look like before?

### After

What will it look like after merging?

### A11y

How do keyboard navigation and screen readers work with this PR?