Set Model ID searchParam on change (#1218)
## Description

Closes: allenai/playground-issues-repo#738

This updates the URL with a searchParam for the model, it adds
`model=XYZ` for single thread, and `model-X=ABC` for comparison.

This does not add searchParams on page load, from a loader.

It very dumbly adds model if the current page has it, is this something
we want, and is it something we want done well, or just forwarding
searchParams an acceptable solution?

## Reviewer Instructions

Anything a review needs to know about this PR to help them give better
feedback. Are there specific questions you have? Are there areas you are
unsure of? Does this come from or relate to another part of the
codebase?

## Testing Plan

Describe what you did to test the change.

## Output / Screenshots

### Before

What did it look like before?

### After

What will it look like after merging?

### A11y

How do keyboard navigation and screen readers work with this PR?