Agent parameters (#1216)
## Description

Closes: allenai/playground-issues-repo#700

I had originally thought maybe we could key off inferenceConstraints,
and keep a unified parameter drawer, but I can see how something like
canCallTools will be problematic (agents can, but its not configurable).

So, I split up the Parameters (and content), to be defined by the page.

The actual values are stored in the inferenceOpts (can change, if
needed).

I opted to split the constraints out, but I could probably put the
agent's constraints into the inferenceConstraints (it would require the
changes for parameters UI depending on the presence of a constraint
object). We could also move constraints to always exist if we want to do
that.

## Reviewer Instructions

Anything a review needs to know about this PR to help them give better
feedback. Are there specific questions you have? Are there areas you are
unsure of? Does this come from or relate to another part of the
codebase?

## Testing Plan

Added an e2e -- and added a skipped test because we don't actually load
the values (and I'm not certain if they are in the right place (use
inferernceOpts, or a new object for agents).

Tests can be updated and expanded when we have schema and end to end
flow.

## Output / Screenshots

<img width="1487" height="1293" alt="image"
src="https://github.com/user-attachments/assets/1d75f745-d4ed-4f67-b284-749f26d52596"
/>
<img width="387" height="677" alt="image"
src="https://github.com/user-attachments/assets/f10a4b8c-a6d5-40ff-bdc4-20bbe6065526"
/>


### A11y

How do keyboard navigation and screen readers work with this PR?