put scopes in the right place for a0 setup (#1233)
@warmbowski @bronzle and I found this when looking at the other auth
issue. Setting this doesn't break anything and ensures we're always
getting/setting the correct scopes for our token. It seems like setting
the scopes on `getToken` only gives the scopes sometimes, which may have
been causing some of the issues with not being able to access the admin
page.

This may simplify some of the code here and on the API that we have for
checking permissions 🤦