visually hidden breaking thread display (#1253)
## Description

I think the visually hidden element is messing up the stack, at least in
my quick debugging.

I moved it and it seems to be OK.

## Reviewer Instructions

Anything a review needs to know about this PR to help them give better
feedback. Are there specific questions you have? Are there areas you are
unsure of? Does this come from or relate to another part of the
codebase?

## Testing Plan

Describe what you did to test the change.

## Output / Screenshots

### Before


https://github.com/user-attachments/assets/31d56022-175a-46bd-ae50-700e249feeae

### After



https://github.com/user-attachments/assets/94cfa63d-5522-4699-bdd9-807c399b3888



### A11y

How do keyboard navigation and screen readers work with this PR?