Don't wrap menu in FileTrigger (#1266)
closes: https://github.com/allenai/playground-issues-repo/issues/811

## Description

Describe the change and its purpose. Include a ticket if applicable.

## Reviewer Instructions

Anything a review needs to know about this PR to help them give better
feedback. Are there specific questions you have? Are there areas you are
unsure of? Does this come from or relate to another part of the
codebase?

## Testing Plan

Describe what you did to test the change.

## Output / Screenshots

### Before


https://github.com/user-attachments/assets/88b3cf53-4392-42fa-a1d9-b3656b59900a


### After




https://github.com/user-attachments/assets/fc0c3bdc-faea-4004-a80e-1fc8c5c2b5fb



### A11y

How do keyboard navigation and screen readers work with this PR?