fix slider scrolling (#1270)
## Description

Fix scrolling in slider. make sure first and last are accessible through
sliding and scrolling.
Add manual scrolling back in (accidentally removed in previous PR).

There was a small refactor to remove the useEffect watching for
scrollIndex changes. the scrolling has moved to a callback.

## Reviewer Instructions

Just check the behavior when sliding to first or last image. It should
stick.

## Testing Plan

Describe what you did to test the change.

## Output / Screenshots

### Before

Scrolling to end could end up in bounce-back

https://github.com/user-attachments/assets/9413edd8-e9a7-473d-84c4-1bac1e269d07


### After

Scrolling is solid now

https://github.com/user-attachments/assets/62878c4b-68f2-44f6-a4e5-b0a0cb3174b2

note the scrollbar at the bottom

https://github.com/user-attachments/assets/2fe5b6c3-a646-47d3-87df-3ef65bdc41e1



### A11y

How do keyboard navigation and screen readers work with this PR?