Ios fix (#1280)
## Description

instead of using the data that is somehow un-reset that is passed to our
handler, we can pass the watched fileslist and pass that up.

This seems to work, though I'd really like to know wat is going on here.

## Reviewer Instructions

Anything a review needs to know about this PR to help them give better
feedback. Are there specific questions you have? Are there areas you are
unsure of? Does this come from or relate to another part of the
codebase?

## Testing Plan

Describe what you did to test the change.

## Output / Screenshots

### Before

What did it look like before?

### After

What will it look like after merging?

### A11y

How do keyboard navigation and screen readers work with this PR?

---------

Co-authored-by: Caleb Ouellette <calebo@allenai.org>