Point picture max image height (#1289)
## Description

This PR reverts previous changes to PointPicture that was incompatible
with image slider. Then it adds a prop to constrain image height. this
still removes any overflow.

previous changes https://github.com/allenai/olmo-ui/pull/1281

## Reviewer Instructions

this requires a thread with a single tall image. The main thing is that
the image works in the thread list AND in the Lightbox (with the
Lightbox slider being responsive with no max height)

## Output / Screenshots

### Before

What did it look like before?

### After



https://github.com/user-attachments/assets/9558285b-d4f5-47a8-bbec-ffb1d04cfe4d



### A11y

How do keyboard navigation and screen readers work with this PR?