Pop Out tracking input (#1287)
## Description

This turns the tracking input into a popout, modal-esque interaction.

I removed minHeight for global videos, and set it specifically on the
Tracking/Counting implementations, and the tracking input is constrained
by the container.

I suspect that a lot of the video player stuff could be consolidated,
but I'm hesitant to do that right before launch.

## Reviewer Instructions

Anything a review needs to know about this PR to help them give better
feedback. Are there specific questions you have? Are there areas you are
unsure of? Does this come from or relate to another part of the
codebase?

## Testing Plan

Describe what you did to test the change.

## Output / Screenshots

### Chrome

https://github.com/user-attachments/assets/2e3f0f7e-27e0-4d82-bb67-e2c2dd520675

#### Firefox

https://github.com/user-attachments/assets/64db3026-847b-4a4d-b9ae-effe46302447

#### Safari


https://github.com/user-attachments/assets/19df6a9e-a122-4f71-9bc4-f2d8c523944b

#### Output


https://github.com/user-attachments/assets/87d98808-442c-4c1a-b067-a724e4566af0


What will it look like after merging?

### A11y

How do keyboard navigation and screen readers work with this PR?