Add validations for content + pointing and files (#1293)
Closes: https://github.com/allenai/playground-issues-repo/issues/849

## Description

Added custom validation for content that passes if there are inputParts.
I also went ahead and added a fileSize validation too.

## Reviewer Instructions

The current implementation does not show a message for the content in JS
-- but does from API. I've kept this behavior, but curious if we want to
change that? (it can be sort of annoying, so maybe thats why?)

## Testing Plan

Describe what you did to test the change.

## Output / Screenshots

### Before

What did it look like before?

### After

What will it look like after merging?

### A11y

How do keyboard navigation and screen readers work with this PR?