add button wrapper for touch events (#1306)
## Description

This adds a touch interaction to pop the disabled tooltip message in the
FileUploadButton on mobile, while allowing existing behavior on desktop.

Note: Popping tooltips on disabled buttons goes against React-aria
opinions. If we do this, it should be rare.

closes https://github.com/allenai/playground-issues-repo/issues/825

## Reviewer Instructions


## Testing Plan

Describe what you did to test the change.

## Output / Screenshots

### Before

No tooltip shown on mobile

### After

mobile and desktop interaction 
* not apparent, but I am "touching" the button in mobile simulators
* I am hovering the button on desktop browser

https://github.com/user-attachments/assets/1644b33c-e38d-4c15-949f-770661f19399


### A11y

A11y should remain as it was before