Validate for filetypes (#1322)
## Description

realized we didn't actually check on front end for the wrong file type
(since we limit the types via file chooser).

One of the prompt templates was showing as invalid because it says
application/octet-stream -- which triggers on the API side (but not the
FE).

This collects the invalid file types and displays a message to the user.

## Reviewer Instructions

Anything a review needs to know about this PR to help them give better
feedback. Are there specific questions you have? Are there areas you are
unsure of? Does this come from or relate to another part of the
codebase?

## Testing Plan

Describe what you did to test the change.

## Output / Screenshots

### Before

What did it look like before?

### After

What will it look like after merging?

### A11y

How do keyboard navigation and screen readers work with this PR?