Input styling tweaks (#1313)
Closes: https://github.com/allenai/playground-issues-repo/issues/894

## Description

Removed the MUI shared button in favor of a RAC one with varnish/panda
styles.

Adds disabled state and styling to the input when the prompt is
submitting.

We removed the abort on navigation at some point, I looked at it and
made a change to abort on any nav that is a separate URL

Not 100% on all the colors focus.

## Testing Plan

Describe what you did to test the change.

## Output / Screenshots



https://github.com/user-attachments/assets/4ca8be3d-4c2b-4e3f-bdd1-d98ae99ea5ec



### A11y

How do keyboard navigation and screen readers work with this PR?