add select_account to login authorizationParams (#1334)
## Description

On playground, when a signed in user logs out and then they log back in
again, they aren't presented with the google screen to select or add an
account to login with.

This only happens when they are logged into a single google account. The
way around this in the current flow is to log into a different google
account outside of playground. Once you have more then one google
account set up, you will be prompted in the Playground login flow to
select or add another google account.

This settings change forces the select or add account screen to always
appear on login, making it easier to switch or add google accounts.

## Reviewer Instructions


## Testing Plan

Describe what you did to test the change.

## Output / Screenshots

### Before



https://github.com/user-attachments/assets/532bdbc1-5c8d-4d9a-8de2-2394647ff8c2



### After



https://github.com/user-attachments/assets/f82fa7e1-9297-4eff-8828-ae4804161cfa



### A11y

How do keyboard navigation and screen readers work with this PR?